# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Taskaya API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, completed, approved, reopened]
        - name: assigneeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFilter
          in: query
          schema:
            type: string
            enum: [today, this_week, this_month, custom]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Employee marks task as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/approve:
    patch:
      operationId: approveTask
      tags: [tasks]
      summary: Manager/deputy approves completed task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/reopen:
    patch:
      operationId: reopenTask
      tags: [tasks]
      summary: Manager/deputy reopens a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task reopened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/reassign-request:
    post:
      operationId: requestReassignment
      tags: [tasks]
      summary: Assignee requests task reassignment to another user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignRequestInput"
      responses:
        "200":
          description: Reassignment requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/reassign-approve:
    patch:
      operationId: approveReassignment
      tags: [tasks]
      summary: Manager approves a pending reassignment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reassignment approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/reassign-reject:
    patch:
      operationId: rejectReassignment
      tags: [tasks]
      summary: Manager rejects a pending reassignment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reassignment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, mobile, role, isActive, mustChangePassword, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [owner, deputy, member]
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        groupId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Task:
      type: object
      required: [id, title, description, assigneeId, creatorId, deadline, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        assignee:
          $ref: "#/components/schemas/User"
        creatorId:
          type: integer
        creator:
          $ref: "#/components/schemas/User"
        deadline:
          type: string
        status:
          type: string
          enum: [open, completed, approved, reopened]
        reassignToId:
          type: ["integer", "null"]
        reassignTo:
          $ref: "#/components/schemas/User"
        reassignStatus:
          type: ["string", "null"]
          enum: [pending, approved, rejected, "null"]
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ReassignRequestInput:
      type: object
      required: [requestedAssigneeId]
      properties:
        requestedAssigneeId:
          type: integer
    TaskInput:
      type: object
      required: [title, description, assigneeId, deadline]
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        deadline:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        deadline:
          type: string
    Attachment:
      type: object
      required: [id, taskId, filename, originalName, mimeType, size, url, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        createdAt:
          type: string
...